Move migration into its own module
